+++ tests/testthat/test-plot_pit.R
@@ -2,7 +2,7 @@ test_that("plot_pit() works as expected with quantile forecasts", {
   pit <- example_quantile %>%
     na.omit() %>%
     as_forecast() %>%
-    pit(by = "model")
+    get_pit(by = "model")
   p <- plot_pit(pit, breaks = seq(0.1, 1, 0.1))
   expect_s3_class(p, "ggplot")
   skip_on_cran()
@@ -13,7 +13,7 @@ test_that("plot_pit() works as expected with integer forecasts", {
   pit <- example_integer %>%
     na.omit() %>%
     as_forecast() %>%
-    pit(by = "model")
+    get_pit(by = "model")
   p <- plot_pit(pit)
   expect_s3_class(p, "ggplot")
   skip_on_cran()